+++ app/Ecommerce/Backend/Controllers/Admin/SubCategory/Controllers/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace Ecommerce\Backend\Controllers\Admin\SubCategory\Controllers;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use Ecommerce\Backend\Controllers\Admin\Category\Models\Category;
+use Ecommerce\Backend\Controllers\Admin\ChildCategory\Models\ChildCategory;
 use Ecommerce\Backend\Controllers\Admin\SubCategory\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -102,6 +103,14 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+
+        if($childCategory > 0){
+            return response(['status' => 'error',
+                'message' => 'This items contain, sub items for delete this
+                you have to delete the sub items first!'
+            ]);
+        }
 
         $subCategory->delete();
